add codes for data lifestime estimation method

+++ src/IOsim_q.c
@@ -162,7 +162,7 @@ block* write_job_start_q(rttask* tasks, int taskidx, int tasknum, meta* metadata
         metadata->tot_write_cnt++;
         metadata->write_cnt_per_cycle[lpa]++;
         metadata->reserved_write++;
-        metadata->avg_update[lpa] = (metadata->avg_update[lpa]*(metadata->write_cnt[lpa]-1)+(cur_cp - metadata->recent_update[lpa]))/(long)metadata->write_cnt[lpa];
+        metadata->avg_update[lpa] = (metadata->avg_update[lpa]*(long)(metadata->write_cnt[lpa]-1)+(cur_cp - metadata->recent_update[lpa]))/(long)metadata->write_cnt[lpa];
         //printf("avg : %ld, recent : %ld\n",metadata->avg_update[lpa],metadata->recent_update[lpa]);
         metadata->recent_update[lpa] = cur_cp;
         req->type = WR;
